dbcheck: add a confirm_all for variable definitions of "all"
This will be used for a gPLink check where (after a domain rename) you
might want to fix all that refer to the old domain, but not the ones
that refer to another domain you rely on. There might be hundreds of
gplinks, and you want to confirm-all of the old-domain ones, and
confirm-none of the other-domain ones.

Signed-off-by: Douglas Bagnall <douglas.bagnall@catalyst.net.nz>
Reviewed-by: Jennifer Sutton <jennifersutton@catalyst.net.nz>