samba-tool dbcheck: --check-gpo-links to fix gplinks after rename
The domain backup system can be used to rename a domain, which works
for almost everything except GPOs. In particular, gPLinks that refer
to the old domain are not modified, but the groupPolicyContainer
objects they refer to are copied to the new domain.

That is, the gPLink will look like
 CN={guid},CN=Policies,CN=System,DC=OLD
when it really means to point to
 CN={guid},CN=Policies,CN=System,DC=NEW

It is possible for gPLinks to refer to other trusted domains, so we
need to be careful in this renaming, which means

1. this test is only done with --check-gpo-links, and
2. the yes/no/all/none dialog is per domain (allowing e.g.
   "all" for DC=OLD, and "none" for DC=TRUSTED).

Signed-off-by: Douglas Bagnall <douglas.bagnall@catalyst.net.nz>
Reviewed-by: Jennifer Sutton <jennifersutton@catalyst.net.nz>