samba-tool dbcheck: --check-gpo-links to fix sysvol after rename
When a backup has been restored under a different name, the sysvol is
not changed, nor are the gPCFileSysPath attributes that refer to it.

A gPCFileSysPath is an UNC that looks like

  \\example.com\sysvol\example.com\Policies\{<gpo guid>}

the first "example.com" is a host reference, the "sysvol" is a service
name, and the second "example.com" is a file system path in the sysvol.

If the domain has been renamed from "example.com" to "example.biz",
the host is probably now wrong and should use the new DNS name. The
file system path should *by convention* also be changed, but the
actual policies will still be there in the "example.com" directory, so
it might be easier to leave them as is. If the gPCFileSysPath is
changed, the file system will need to me modified accordingly. In
dbcheck we don't actually know where the sysvol is, so we can't
attempt that here.

Thus there are separate questions for both parts.

Signed-off-by: Douglas Bagnall <douglas.bagnall@catalyst.net.nz>
Reviewed-by: Jennifer Sutton <jennifersutton@catalyst.net.nz>