samba-tool dbcheck: reduce undocumented magic in --yes
If --yes is followed by a string that doesn't look like another
command line option, that string is interpreted by --yes as a space
separated list of confirmation options to be treated as the --yes only
applied to them (the list of confirmation options is only available by
reading ../dbchecker.py).

One problem with this system is that dbcheck can also take a DN as a
non-option, as per USAGE:

    Usage: samba-tool dbcheck [<DN>] [options]

and if someone went `samba-tool dbcheck --fix --yes DC=example...`,
it would not have worked as expected (though the damage would be
restricted to confusion). On the other hand,
`samba-tool dbcheck --yes --fix DC=example...` would be fine.

This usage of --yes is only seen in blackbox tests, and has never been
mentioned in WHATSNEW or public lists, so we replace it with a new rule
--yes-restricted-to-rule. Unlike the --yes='a b' form that split on
whitespace, the new rule must be specified multiple times to say yes to
more than one rule, like this:

samba-tool dbcheck --quiet --fix                              \
   --yes-restricted-to-rule=fix_all_missing_backlinks         \
   --yes-restricted-to-rule=remove_plausible_deleted_DN_links

It remains hidden from --help for now, because we don't document anywhere
what the available switches are.

Signed-off-by: Douglas Bagnall <douglas.bagnall@catalyst.net.nz>
Reviewed-by: Jennifer Sutton <jennifersutton@catalyst.net.nz>